correções aplicadas

+++ Exercícios/Ex6/DoacaoDeSangue.java
@@ -7,10 +7,10 @@ public class DoacaoDeSangue {
         //6 - Para doar sangue é necessário ter entre 18 e 67 anos. Faça um aplicativo que pergunte a idade de uma pessoa e diga se ela pode doar sangue ou não.
 
         Scanner scanner = new Scanner(System.in);
-            int idade;
+            short idade;
 
             System.out.println("Qual sua idade?");
-            idade = scanner.nextInt();
+            idade = scanner.nextShort();
 
             if(idade >= 18 & idade <= 67){
                 System.out.print("Você pode doar sangue");
